+++ src/app/api_new/companies/get_by_query/route.tsx
@@ -1,5 +1,6 @@
 import { NextResponse } from "next/server";
 import clientPromise from "@/app/lib/mongodb";
+export const dynamic = "force-dynamic";
 
 interface ICompanyQuery {
   _id: string;
