+++ src/app/api_new/prospects/get_all_my_prospects/route.tsx
@@ -1,5 +1,6 @@
 import { NextResponse } from "next/server";
 import clientPromise from "@/app/lib/mongodb";
+export const dynamic = "force-dynamic";
 
 export async function GET(request: Request) {
   try {
